Add more tests and split up one file into modules
