+++ src/pizza/business.py
@@ -1,16 +1,6 @@
-import random
 from collections import defaultdict
-import sys
-import time
 from collections import UserDict
 from typing import Self
-
-import click  # but Typer is a bit cleaner
-
-
-@click.group()
-def cli():
-    pass
 
 
 class FoodItem:
@@ -22,6 +12,8 @@ class FoodItem:
 
 
 class PizzaRecipe(FoodItem):
+    """Specialization of food - pizza"""
+
     type_of_food = "pizza"
 
     def __init__(self, size="L") -> None:
@@ -36,7 +28,7 @@ class PizzaRecipe(FoodItem):
     @classmethod
     @property
     def name(cls) -> str:
-        return cls.__class__.__name__.capitalize()
+        return cls.__mro__[0].__name__.capitalize()
 
     @classmethod
     @property
@@ -65,7 +57,7 @@ class PizzaRecipe(FoodItem):
 
 
 class Restaurant:
-    def __init__(self, menu: "CustomDict") -> None:
+    def __init__(self, menu: "DictAssortment") -> None:
         self.menu = menu
         self.stock = defaultdict(list)
 
@@ -89,7 +81,6 @@ class Restaurant:
         food = self.retrieve_from_stock(client)
         print("picked up from restaurant")
         return food
-        # client.add_to_stock(food)
 
     def make_order(self, pizza_name, client: "Client", is_delivery=False):
         print("was made a request for pizza")
@@ -98,7 +89,6 @@ class Restaurant:
         self.add_to_stock(client, pizza)
         if is_delivery:
             self.deliver(client)
-        # return pizza
 
     def deliver(self, client: "Client") -> None:
         # do some work with timer
@@ -141,32 +131,7 @@ class Client:
         return food
 
 
-def add_latency(fn):
-    def wrapper(*args, **kwargs):
-        time.sleep(random.randint(1, 3) / 5)
-        result = fn(*args, **kwargs)
-        return result
-
-    # time.sleep(0)
-    return wrapper
-
-
-def log(str_template: str):
-    def outer_wrapper(fn):
-        def wrapper(*args, **kwargs):
-            time_start = time.time()
-            # @wraps(fn) # apply wrapper from functools
-            result = fn(*args, **kwargs)
-            lapsed_time = time.time() - time_start
-            print(str_template.format(lapsed_time))  # apply to placeholder
-            return result
-
-        return wrapper
-
-    return outer_wrapper
-
-
-class CustomDict(UserDict):
+class DictAssortment(UserDict):
     def __init__(self):
         super().__init__()
 
@@ -181,7 +146,7 @@ class CustomDict(UserDict):
         return super().__contains__(item.lower())
 
 
-assortment = CustomDict()
+assortment = DictAssortment()
 
 
 def pizza_to_assortment(cls):
@@ -207,62 +172,9 @@ class Hawaiian(PizzaRecipe):
     emoji = "🍍"
 
 
-# menu_str = ""
-# for k, v in assortment.items():
-#     menu_str += f"- {k} {v.emoji} : {v.clean_recipe}\n"
 menu_str = "\n".join(
     f"- {v.name} {v.emoji} : {v.clean_recipe}" for v in assortment.values()
 )
-
-
-@cli.command()
-def menu():
-    print(menu_str)
-
-
-class EmojiSelector:
-    # select emoji by name from list
-    def __init__(self):
-        pass
-
-    def __call__(self, *args, **kwargs):
-        pass
-
-
-@cli.command()
-@click.option("--delivery", default=False, is_flag=True)
-@click.argument("pizza", nargs=1)
-def order(pizza: str, delivery: bool):
-    pizza = pizza.capitalize()
-    if pizza not in assortment:
-        print("No such pizza in the assortment, here's the menu:")
-        print(menu_str)
-        sys.exit()
-    ordered_pizza = assortment[pizza]()
-    print("I want to order", pizza, ordered_pizza.emoji)
-    bake(ordered_pizza)
-    if delivery:
-        deliver(ordered_pizza)
-    else:
-        pick_up(ordered_pizza)
-
-
-@log("Picking up took {:.2f} seconds")
-@add_latency
-def pick_up(pizza):
-    print("🏎️ picked up ", end="")
-
-
-@log("Cooking took {:.2f} seconds")
-@add_latency
-def bake(pizza):
-    print("👩‍🍳baked ", end="")
-
-
-@log("Delivery took {:.2f} seconds")
-@add_latency
-def deliver(pizza):
-    print("🛵 delivered ", end="")
 
 
 if __name__ == "__main__":
