+++ src/pizza/emoji.py
@@ -1,0 +1,7 @@
+class EmojiSelector:
+    # select emoji by name from list
+    def __init__(self):
+        pass
+
+    def __call__(self, *args, **kwargs):
+        pass
