--- tests/test_main.py
@@ -1,5 +1,0 @@
-import pizza.main_pizza as m
-
-
-def test_pizza():
-    assert m.Margherita() == m.Margherita()
